Add files via upload
The goal of the project was to perform an explanatory data analysis on data set of the auto claims filed by customers of the automobile insurance company in order to identify the factors impacting the profitability of the company. The prime objective of the analysis was to look into the segments of the company’s most and the least profitable customers for any clear patterns and indications which could potentially help the company minimize losses which in turn would increase the revenues. A set of 8 questions was designed and answered so as to maximize the value of the insights extracted. The questions explored how customer demography, channel of sales used to onboard the customer, customer’s class of vehicle and customer gender impacted the revenue generated. The questions also investigated the influence of insurance policy and coverage type on the net revenue of the company. 
1.	Are there types of customers, based on their domiciled state, that are highly profitable?
2.	Do certain policies have a lower number of claims, leading to large profits?
3.	Does "Sales channel" utilized to on-board customers have an impact on the number of claims?
4.	Is there an ideal subscription length (months) which ensures maximum revenue for the company?
5.	How does the class of vehicle impact the customer lifetime revenue generated for the company?
6.	What kind of coverage had the greatest positive impact on the company's revenue generation?
7.	How do Customer Gender and Residential Demography influence company's profitability?
8.	How does customer Gender and Monthly Premium impact company revenues?
Key Findings
A lot of interesting observations were made while analysing the data set. The customers based out of the states of California, Oregon and Arizona proved to be the most profitable of all the other states. Knowing this could help the company run targeted marketing campaigns in the areas where customer base generates maximum revenues for the company. The customers on the Personal Policy had the highest number of claims. This insight could help the company decide which policy to penalize or incentivize in order to maximize the revenues. The Agent attracted customers had the lowest claims on average (2.38 per customer) while the customers who signed up using the website had the highest claims on average (2.41 per customer). This knowledge can be of great value to the company’s marketing and sales team when choosing which sales channel to exhaust the most. The customers with active policies in between 40 to 55 months turned out to be highly profitable in terms of revenue generated. The company's highly profitable clients (the customers who yielded net revenues of more than the average revenue per customer (> $922.1)) had their policies active for an average of 46.42 months. This awareness could encourage the company to launch targeted advertisement campaigns to attract long term clients. The Luxury Suvs, Luxury Cars, and SUVs generated highest average lifetime revenues for the company. Knowing which type of vehicle earns the company the most money could help the decision makers make informed decisions. The customers on the Extended and Premium coverage have most of the data points above the zero mark (below which the company starts incurring losses). On the contrary, the Lifetime revenues for the Basic coverage are evenly distributed around the value of zero indicating a great number of customers on the coverage generated net revenues of less than zero for the company. This information could prove to be of great value when designing the company’s loss mitigation plan. Suburban Women grossed the company $2,531,073 in profits which is way higher than that of any other demographic or gender. This discovery could help the company run highly tailored offers to attract more clients from the same demography. The company had to pay a grand total of $1,213,122 to their Male clients in claims while they had to spend a far lesser amount of $758,704 covering up their Female Customer claims. These stats could help the company set insurance premiums in order to minimize the losses. 
Recommendations
The clientele from the states of California and Oregon made the company $189,870,6 and $167,007,8 in revenues respectively which is significantly higher than that of any other state. Based on this observation, the company should run massive advertisement campaigns targeting these two states. This would enhance the likelihood of company acquiring customers that would prove highly profitable.
A total of 11131 claims were filed by the customers holding the Personal insurance policy while significantly lower number of claims were received for customers on the Corporate (3175 claims) and Special plans (640). Considering the stats, the company should look to run B2B marketing campaigns and offer Employer/Affinity group discounts in order to attract corporate clients which could potentially boost the net revenue of the company.
An in-depth analysis was performed on the filtered data on each of the marketing channels in order to determine average claims filed by customers on-boarded by the particular channel of sales. The data suggests that the Agent attracted customers had the lowest claims on average (2.38 per customer) while the customers who signed up using the website had the highest claims on average (2.41 per customer). Based upon this information, the company should look to prioritize agent marketing over any other sales channel. This could ensure that the company wins over clients who are less likely to file more claims which in turn will enhance revenues. 
The company should offer special discounts exclusively to customers who have been with them for over 3 years. This is based on the fact that the customers with active policies in between 40 to 55 months turned out to be highly profitable in terms of revenue generated. The company's highly profitable clients had their policies active for an average of 46.42 months. The special discounts on customer loyalty would help the company with the client retention in turn leading to increased revenues. 
Luxury SUVs, Luxury Cars, and Sports Cars generated lifetime revenues of $6,381, $5,669 and 2,158 on average respectively while the typical two-door and four-door cars netted $269 and $271 in revenues on average respectively. On the basis of this analysis, the company should work to diversity in order to attract clients owning Luxury and Sports cars. The company should offer discounts and lower premiums on these cars so as to become the client’s number one choice in this niche. This would help the company see increased revenues in the long term. 
The average Customer Lifetime Value of customers on the Basic, Extended and Premium coverage came out to be $326, $1,538 and, $2,879 respectively. The company should encourage more and more customers to opt for the Extended and Premium coverage by incentivizing those plans. 
 Of all the customers generating more than average revenue for the company, the Suburban and Urban Women helped the company make $2,531,073 and $2,011,879 respectively as compared to their male counterparts ($552,967 and $ 815,521 respectively). The company had to pay a grand total of $1,213,122 to their male clients in claims while they had to spend a far lesser amount of $758,704 covering up their Female customer claims. Based upon these facts the company should work to execute highly targeted campaigns to attract Suburban and Urban women. The company could also offer lower premiums to females from these demographics which in turn could skyrocket the company’s revenues.